Add support for RFC ChaCha20 (used in ChaCha20Poly1305 RFC).
This replaces the AVX implementation from openssl__chacha20_poly1305_cf.patch
with the implementation in
openssl__chacha20_poly1305_draft_and_rfc_ossl102g.patch.

It adds NewRFC and NewDraft. NewDraft behavoir identical to the previous New,
and NewRFC which uses a 96 bit nonce and a 32 bit counter.

It adds RFCNonceSize and DraftNonceSize. Previous, and current, NonceSize is
equal to DraftNonceSize (64 bits). RFCNonceSize (96 bits) is required for
NewRFC.

New is updated to select NewRFC or NewDraft depending on len(nonce).

Performance is identical.

The pure Golang implementation from codahale/chacha20 is included as
internal/ref and has been expanded to include support for RFC. XChaCha support
(which is included in codahale/chacha20) has been removed.